+++ spec/features/index_page/index_page_spec.rb
@@ -1,5 +1,6 @@
 RSpec.describe '/content' do
   include GdsApi::TestHelpers::ContentDataApi
+  include RequestStubs
   include TableDataSpecHelpers
   let(:metrics) { %w[pviews upviews searches feedex words pdf_count satisfaction useful_yes useful_no] }
   let(:items) do
